+++ surpyval/univariate/parametric/__init__.py
@@ -48,6 +48,7 @@ from .distributions import (
 from .mixture_model import MixtureModel
 from .parametric import Parametric
 from .parametric_fitter import ParametricFitter
+from .royston_parmar import RoystonParmar, RoystonParmarModel
 
 
 class NeverOccurs(Distribution):